k8s: factorio: stable tag, backup saves on start
Now tracking the stable tag so the server updates with whatever is out
on steam.

Since updating the server could break saves, and definitely stop
downgrading from being easy and safe, the save backup PVC was added.
Before the server starts the current contents of /factorio/saves is
copied to the factorio-saves-backups PVC with a date time component in
its path.